requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
discord.py==1.7.3
idna==3.3
multidict==6.0.2
numpy==1.21.5
praw==7.5.0
prawcore==2.3.0
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.2.3
wikipedia==1.4.0
yarl==1.7.2
